Fix: Type error in pdfUtils.ts
The code was producing a type error in `pdfUtils.ts` at line 21, column 20.  This commit addresses the type error.

+++ src/utils/pdfUtils.ts
@@ -17,6 +17,6 @@ export const mergePDFs = async (pdfBlobs: Blob[]): Promise<Blob> => {
     await merger.add(arrayBuffer);
   }
 
-  const mergedUint8Array = await merger.save('Uint8Array');
+  const mergedUint8Array = await merger.save('Uint8Array') as Uint8Array;
   return new Blob([mergedUint8Array], { type: 'application/pdf' });
 };
